[CISCO] PT 6.4.1 - Implement EtherChannel
Completed the following tasks:
You have been tasked with designing an EtherChannel implementation for a
company that wants to improve the performance of the switch trunk links.
You will try several different ways of implementing the EtherChannel
links in order to evaluate which is the best for the company. You will
build the topology, configure trunk ports and implement LACP and PAgP
EtherChannels.