[CISCO] PT 7.4.1 - Implement DHCPv4
Completed the following task:
As the network technician for your company, you are tasked with
configuring a Cisco router as a DHCP server to provide dynamic
allocation of addresses to clients on the network. You are also required
to configure the edge router as a DHCP client so that it receives an IP
address from the ISP network. Because the server is centralized, you
will need to configure the two LAN routers to relay DHCP traffic between
the LANs and the router that is serving as the DHCP server.